A series of fixes to allow "eups distrib create" to work once more
1/ Correctly create and use generic and non-generic manifests (via --use-flavor flag)
This required me to use Distrib.flavor not Eups.flavor when doing instally things,
as an Eups.flavor of "generic" breaks the world.
2/ Made warnings about table files near-but-not-on buildpath work once more; supported
"*" in build path elements to mean, "Search Recursively"
3/ Reinstated code to automatically overwrite files for top-level product
4/ Print name of log file so users can find it.
5/ Added a class, BuildfilePatchCallbacks, to allow users to define callbacks
to rewrite tagnames and map them to svn paths in buildfiles.  Implemented the standard
rewrite of tags/svnXXX --> trunk -r XXX as a callback, and added the LSST specific
ticket tagnames to eupsLsstStartup.py
6/ Fixed some of the Eups-instance local current code, and declare products
in top-level table file (locally) current so that they'll be used when 
generating manifests for dependent products [I hope; not well tested, but
it seems to work]
7/ Made eups.declare() handle case that it was passed a productDir of /dev/null
and failed to find its true value (message + treat as "none")
8/ Use setup --keep when adding products to build files, or setting up . in
the build file itself
9/ Handle rootdir == "/dev/null" when looking for tables
10/ In eups itself, process common arguments such as -f before parsing
command-specific arguments.  This permits e.g. "eups -f Darwin distrib -f create ..."
11/ Made distrib -f work; it'd been setting and ignoring a flag
12/ Factored processing of pkgroots into getPkgroots, and use it to support
eups admin clearServerCache 
13/ Once more search for table files locally, not just via server.  Useful when
trying to _create_ a distribution...
14/ Make --incomplete work again
15/ Make clearConfigCache actually work
16/ Validate config files
17/ Use range(len(locations)) not xrange(0, 3) for robustness
18/ Handle versions being Current (or Stable?) --- classes, not strings.  This
is a hack; I should find a better solution.
19/ Fix logic checking permissions to write cache files