(Working over the Pacific)
Fixed --keep to work with the VRO.  Code is simpler than before (I just pushed pseudo-tag "keep" onto the
VRO); there are some possibly-fossil remnants of the old solution.

Support the use of a file as a "tag"; the file can be a (restricted) table file, or the output of eups list
(i.e. columns of product and version and maybe other stuff).

Add tag to start of VRO if no VRO is specified for the tag

Initialise all lists of tags as lists, not strings, so the user can append without worrying about whitespace

Don't let product foo in a tablefile override an explicit version specified on the command line

Protect against people using internal tags such as "keep" by mistake