Permit values containing the delimiter to be added to paths
It probably isn't a good idea to use this, but it shouldn't be forbidden.  A table file like

pathPrepend(FOOVAL, A)
pathPrepend(FOOVAL, B)
pathPrepend(FOOPATH, ${FOOVAL})
pathPrepend(GOOPATH, ${FOOVAL})

should work, although it _won't_ unsetup properly.  It'd be better to prepend A twice, once to
FOOPATH and once to GOOPATH