eupspkg.sh: support VERSION_PREFIX and VERSION_SUFFIX
$VERSION_PREFIX and $VERSION_SUFFIX are the prefix/suffix that
version_to_gitrev() removes from version before interpreting it as a git
rev (to be interpretable by git-rev-parse).

This lets one have a such as:

	hsc_1.2.3_c

that is converted into a 1.2.3 rev (which will, most likely, be a tag).

Also, when in DEVMODE autoversion() will prepend/append the _PREFIX, _SUFFIX
to the autogenerated version.

Environment variables EUPSPKG_VERSION_PREFIX and EUPSPKG_VERSION_SUFFIX will
be consulted if VERSION_{PREFIX|SUFFIX} are left unspecified.