If _EUPS_ASSUME_CACHES_UP_TO_DATE=1, assume all caches are up-to-date
This is a hack to enable users that know that the caches are not stale
(e.g., because they fully control the access to EUPS_PATH) to say so, thus
skipping many fstat() calls to verify cache freshness.

Implemented primarily to introduce a mechanism to speed up continuous
integration builds with stacks with thousands products.