pkgautoversion: only match conforming tags passed via cmdline
Previous behavior was to immediately return as the version any annotated tag
that was passed as REF on the command line.  This works as expected as long
as one follows a convention to use annotated tags for (numerical) version
only.  The new behavior is to immediately return the tag *only* it it
matches the TAG_PATTERN.

E.g., in LSST, this helps when one run an lsstsw-driven rebuild with a
command line such as:

	rebuild -r w.2015.33 -r 2.0.1 lsst_sims

The former behavior would lead to many packages having a version w.2015.33.
Such version is generally hard to parse (and order), and also leads to a
rebuild of a package that has already been built.  The new behavior will
compute the version using the fall-back rules, as if the tag wasn't there.