Local version string must not include spaces
For a local setup eups creates a version string of the form

    LOCAL:/working/directory

The rest of EUPS assumes that the version number does not contain
spaces and uses space as a delimiter when decoding the contents
of some environment variables.

Spaces in paths are a problem. This change provides encoding
and decoding methods that ensure that spaces are not present
in the final version string but which also allow the original
path to be reconstructed.