eupspkg.sh: attempt to raise ulimit -u to the hard limit
In configurations with many cores, we can sometimes run into a limit on the
number of processes (ulimit -u).  When this happens, the build can terminate
with hard-to-trace errors (e.g., see
https://github.com/RobertLuptonTheGood/eups/issues/58.  We'll try to defend
agains this by pre-emptively raising the limit as far as we can (to the hard
limit).  As changing ulimits may be administratively prohibited, we won't
treat failure as a fatal error.