server.py remove confusing unused code
server.py's remapEntries function had a for loop with an "or" clause
that could not run and was confusing. Removed the "or" clause, simplying
the code with no change in behavior. It appears that the or clause was intended
to support glob expressions in manifest.remap files, so updated the doc string
for remapEntries accordingly.