Fix misuse of utils.commonprefix by replacing with simpler utils.isSubpath
At some point os.path.commonprefix was replaced with utils.commonprefix
(because os.path.commonprefix has a known misfeature that it works
character by character, and so can give unexpected and undesired results).
Unfortunately utils.commonprefix raised an exception when asked to compare
relative and absolute paths, which eups was not expecting.
Even os.path.commonprefix didn't do a nice job of comparing relative
and absolute paths, but simply returned "" to indicate no common path.

It turns out utils.commonprefix was only being used to test if one path
was a subpath of another, which is much easier to test
(and in a way that correctly handles a mix of relative and absolute paths),
so I replaced utils.commonprefix with a new function utils.isSubpath
and modified the calling code accordingly.