Allow redeclaring only of products with 'tag:XXX' versions
Fix a problem introduced in 6e234b40. I think the intent of that commit was
to allow re-declaration of a product's directory if it's been previously
declared with a 'tag:<tagname>' version.  This allows for convenient usage
such as:

    setup -r . -t mytest
    ... change to a different copy of the product ...
    setup -r . -t mytest

where now the second copy is the one that's declared (and declared with a
version of 'tag:mytest', and a tag 'mytest').  The way this was implemented,
however, has had an odd side-effect of allowing things such as:

    eups declare -r /stack/foo/1.2 foo 1.2 -t bar
    eups declare -r /stack/foo/1.4 foo 1.2 -t bar

which would result in the following declared products:

    foo, with version 1.2 in /stack/foo/1.2
    foo, with version tag:bar in /stack/foo/1.4

(and, more broadly, redeclaration with _any_ change would always end up
declaring a product with version 'tag:<tagname>')

The implementation has now been changed to return to the old behavior of
refusing to redeclare if there are changes in productDir, etc., unless the
version equals 'tag:<tagname>'.