Added support for external dependencies in table files
Products marked as --external in setup{Optional,Required} statements are
usually ignored;  the exception is that if you call Table.dependencies() with
listExternalDependencies=True, only the external products are returned.

This is designed to be used by packaging systems that play nicely with eups,
but handle some packages (e.g. flex, bison) differently;  an example is LSST's
conda binary distributions.

N.b. The return value of Action.ProcessArgs() was modified to mostly return the
same arguments, except that the last two (now three) are packed into a dict.