Enhance content of the deprecation message
The deprecation message in its current form makes no obvious
statement as to why the API was being used or where it was being
misused.

With this commit it now reports the file, line number and function
of the caller routine and gives the reason for triggering the
deprecation message.