Use a utility function to check for strings
Sometimes we can end up with unicode strings if the database
has been created on Python 3 but accessed on Python 2. This
causes some strange behavior as many of the checks for
strings were assuming only "str". To be consistent across
versions those checks are now different on Python 2 and Python
3. On Python 3 "str" is the only option but on Python 2 the
comparison is now with "basestring".