Handle "generic" flavour in eups distrib tags files
We don't support multiple flavours properly in class TaggedProductList,
so it's hard to do this right.  The solution is to handle the current
flavor when reading .tags files (mapping "generic" to that flavour), but
not allowing "eups distrib declare" to accept a --filter argument as we
have no way to update multiple flavours.