Fix eups distrib list to obey --tag and --flavor
Note that there isn't enough information in any single file on the
server to tell us which products are available for which flavor, so
to get accurate listings we have to do another round-trip query for
each product/version/flavor.  This is slow, so is only implemented
if you specify the --precise flag.