Add CXX_CXX14_FLAG
Existing code in the eupspkg.sh script currently bundled here
already sniffs the C++ compiler and provides variable
CXX_CXX11_FLAG; this is set to hold the flags, if any, to be passed
to the compiler from an eupspkg package script in order to enable
the closest available approximation to C++11.

Some eupspkg-built packages now wish to move beyond C++11 to C++14.
This commit adds a new CXX_CXX14_FLAG in eupspkg.sh which may be
used for this purpose, maintaining backward compatibility for
eupspkg package scripts that may already be using the existing
CXX_CXX11_FLAG.