Add rel complement (set diff) arg to eups list
This commit adds a new argument to eups list: -C or  --diff. This
argument calculates the relative complement (aka, set difference) of
declared version products with respect to currently setup products.

All dependent products associated with the named product will be used as
a reference, relative to the list of setup products managed by EUPS.