Change "diff" to "overrides"
Following feedback from Nate Lust, the name "diff" may prove to be
confusing to some, as it implies a diff between two versions and not
necessarily a diff of all the product dependencies that come when
setting up that product. It was felt that a name change to "overrides"
(or, "-o" in short) would more accurately describe the operation here -
i.e., it shows what products the end-user has locally set up which
override those they would get when setting up lsst_distrib with a given
version/tag. To further accentuate that difference, dependent products
are now identified by virtue of a leading "|" symbol, as EUPS utilizes
for the "--dependencies" argument.